requirements.txt
######## For Text2Sql ############
allennlp==0.9.0
dill==0.3.1.1
mysql-connector-python==8.0.19
pandas==0.25.3
sqlparse==0.3.0
parsimonious>=0.8.0


######## For Drop ############
allennlp-semparse
allennlp==1.0
nltk
sqlparse>=0.2.4